Move ParquetRecordWriterFactory to its own class and delete ParquetRecordWriter
