+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/job/runner/rdd/SingleFileWritingIterator.java
@@ -38,7 +38,7 @@ import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.ByteArrayType;
 import sleeper.core.schema.type.ListType;
 import sleeper.core.schema.type.MapType;
-import sleeper.io.parquet.record.ParquetRecordWriter;
+import sleeper.io.parquet.record.ParquetRecordWriterFactory;
 import sleeper.sketches.Sketches;
 import sleeper.sketches.s3.SketchesSerDeToS3;
 
@@ -201,7 +201,7 @@ public class SingleFileWritingIterator implements Iterator<Row> {
                 + "/" + outputFilename + ".parquet";
 
         LOGGER.info("Creating writer for partition {} to path {}", partitionId, path);
-        return ParquetRecordWriter.ParquetRecordWriterFactory.createParquetRecordWriter(new Path(path), tableProperties, conf);
+        return ParquetRecordWriterFactory.createParquetRecordWriter(new Path(path), tableProperties, conf);
     }
 
     private String getPartitionId(Row row) {
